requirements.txt
anthropic>=0.36.0
google-generativeai>=0.7.2
pypdf>=3.17.0
openai>=1.0.0
watchdog>=3.0.0
